chore: remove legacy analytics and badge components
